feat: small refactor to move some logic out of the index files and move to using `.args()` and `.returns()` as it's cleaner Signed-off-by: benjdlambert <ben@blam.sh>
