+++ plugins/scaffolder-backend/src/lib/templating/filters/parseEntityRef/index.ts
@@ -13,49 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { parseEntityRef as filter } from '@backstage/catalog-model';
-import { createTemplateFilter } from '@backstage/plugin-scaffolder-node/alpha';
-import { examples } from './examples';
-
-export const parseEntityRef = createTemplateFilter({
-  id: 'parseEntityRef',
-  description:
-    'Extracts the parts of an entity reference, such as the kind, namespace, and name.',
-  schema: z =>
-    z.function(
-      z.tuple([
-        z.union([
-          z.string().describe('compact entity reference'),
-          z
-            .object({
-              kind: z.string().optional(),
-              namespace: z.string().optional(),
-              name: z.string(),
-            })
-            .describe('`CompoundEntityRef`'),
-        ]),
-        z
-          .object({
-            defaultKind: z
-              .string()
-              .describe('The default kind, if none is given in the reference'),
-            defaultNamespace: z
-              .string()
-              .describe(
-                'The default namespace, if none is given in the reference',
-              ),
-          })
-          .partial()
-          .optional(),
-      ]),
-      z
-        .object({
-          kind: z.string(),
-          namespace: z.string(),
-          name: z.string(),
-        })
-        .describe('`CompoundEntityRef`'),
-    ),
-  examples,
-  filter,
-});
+export { parseEntityRef } from './filter';
