+++ plugins/scaffolder-backend/src/lib/templating/filters/pick/filter.ts
@@ -13,4 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { createProjectSlug } from './filter';
+import { createTemplateFilter } from '@backstage/plugin-scaffolder-node/alpha';
+import { get as filter } from 'lodash';
+import { examples } from './examples';
+
+export const pick = createTemplateFilter({
+  id: 'pick',
+  description:
+    'Selects a specific property (e.g. kind, namespace, name) from an object.',
+  schema: z =>
+    z
+      .function()
+      .args(z.any(), z.string().describe('Property'))
+      .returns(z.any().describe('Selected property')),
+  examples,
+  filter,
+});
